requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
dj-stripe==2.6.2
Django==4.1
django-allauth==0.51.0
django-bootstrap4==22.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-rest-passwordreset==1.2.1
django-sesame==3.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
